Update README.md
The main objective of the Automated VAPT Reporting Tool is to automate and streamline the process of vulnerability assessment and penetration testing (VAPT) by integrating Nmap for network scanning and OpenVAS (GVM) for vulnerability assessment, and generating detailed PDF reports in tabular form for easy analysis. The tool aims to enhance the efficiency and effectiveness of VAPT workflows by automating repetitive tasks, providing accurate and comprehensive reports, and facilitating the identification and mitigation of security vulnerabilities.