Create scan.py
This script defines a function called scan_website, which takes a target parameter representing the
website or IP address to be scanned. Within the function, an instance of nmap.PortScanner() is created to
facilitate the scanning process. The nm.scan() method is then called with the specified target and
additional arguments, which in this case are set to -p 80,443 to focus the scan on these ports.The script
proceeds to iterate over the hosts found in the scan results using nm.all_hosts(). For each host, it checks if
the state is 'up' (nm[host].state() == 'up') to ensure that the host is responsive. If the host is up, it prints the
host's IP address (Host: <IP address>).Next, the script iterates over the protocols available for the host
using nm[host].all_protocols(). Within each protocol, it retrieves the open ports using
nm[host][proto].keys(). For each open port, it checks if the state is 'open' (state == 'open'). If the port is
open, it prints the port number (Open Port: <port>).To use the script, you need to provide the target
website or IP address by assigning it to the target variable. Finally, calling the scan_website(target)
function initiates the scan and displays the scan results, specifically the open ports and the IP address of
the target website.